requirements.txt
# InstantNeo Course - Dependencies

# Core library
instantneo[groq]

# Environment variables management
python-dotenv

